chore: Update models and configurations
This commit updates the models and configurations in the codebase. It adds the `description` field to the `Client`, `Flow`, and `Connection` models. It also updates the `Currency` model to include the `rate` and `code` fields. Additionally, the `evolution.php` configuration file is modified to use the `evolution_api:8080` URL. Finally, the `ConnectionFlow`, `ConnectionSetting`, `FlowSession`, `ObserverServiceProvider`, `CreateConnectionRequest`, `UserSeeder`, `MessengerServiceInterface`, and `Message` models are also modified.

Please note that this commit message does not include any meta information such as issue references, tags, or author names.

+++ app/Services/Messenger/Provinder/WhatsappProvinder.php
@@ -2,7 +2,11 @@
 
 namespace App\Services\Messenger\Provinder;
 
+use App\Repositories\Connection\ConnectionRepository;
+use App\Repositories\Message\MessageRepository;
 use App\Services\Messenger\MessengerServiceInterface;
+use Illuminate\Support\Arr;
+use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Http;
 use Illuminate\Support\Facades\Log;
@@ -18,32 +22,47 @@ class WhatsappProvinder implements MessengerServiceInterface
 
     private mixed $callback_url;
 
+    private MessageRepository $messageRepository;
+
+    private ConnectionRepository $connectionRepository;
+
+    private $connectionExists;
+
     public function __construct()
     {
         $this->url = Config::get('evolution.url');
         $this->key = Config::get('evolution.key');
         $this->callback_url = Config::get('app.url');
 
+        $this->messageRepository = App::make(MessageRepository::class);
+        $this->connectionRepository = App::make(ConnectionRepository::class);
+
         $this->request = Http::withHeaders([
             'apikey' => $this->key,
         ])
-        ->acceptJson();
+            ->timeout(60)
+            ->acceptJson();
 
     }
 
     public function createConnection(array|object $data): array|object
     {
+        $instanceName = Str::uuid()->toString();
+        $token = Str::uuid()->toString();
+        $number = Arr::get($data, 'connection_key');
+        $webhook = "{$this->callback_url}/api/integration/whatsapp/callback";
+        $events = ['QRCODE_UPDATED', 'MESSAGES_UPSERT'];
+        $name = Arr::get($data, 'name', 'Conexão padrão');
+        $description = Arr::get($data, 'description', "Nova conexão com o número $number.");
+
         $payload = [
-            'instanceName' => Str::uuid()->toString(),
-            'token' => Str::uuid()->toString(),
+            'instanceName' => $instanceName,
+            'token' => $token,
             'qrcode' => true,
-            'number' => $data['connection_key'],
-            'webhook' => "{$this->callback_url}/api/integration/whatsapp/callback",
+            'number' => $number,
+            'webhook' => $webhook,
             'webhook_by_events' => false,
-            'events' => [
-                'QRCODE_UPDATED',
-                'MESSAGES_UPSERT',
-            ],
+            'events' => $events,
         ];
 
         Log::debug(__CLASS__.'.'.__FUNCTION__.' => start', [
@@ -52,12 +71,79 @@ class WhatsappProvinder implements MessengerServiceInterface
         ]);
 
         try {
-            $response = $this->request->post("{$this->url}/instance/create", $payload);
+            if (! $this->connectionRepository->exists(column: 'connection_key', value: $number)) {
+                $response = $this->request->post("{$this->url}/instance/create", $payload);
 
+                if ($response->successful()) {
+                    $response = $response->json();
+                    $instance = Arr::get($response, 'instance.instanceName');
 
+                    if (! $this->connectionRepository->exists(column: 'token', value: $instance)) {
+                        $createConnection = $this->connectionRepository->create([
+                            'user_id' => rand(1, 5),
+                            'name' => $name,
+                            'description' => $description,
+                            'type' => 'whatsapp',
+                            'connection_key' => $number,
+                            'token' => $instance,
+                            'country' => 'BR',
+                            'is_active' => 0,
+                            'payload' => json_encode($response),
+                        ]);
+
+                        return (object) [
+                            'data' => [
+                                'success' => true,
+                                'message' => 'Nova conexão criada com sucesso.',
+                                'payload' => $createConnection,
+                            ],
+                        ];
+                    }
+                }
+            }
 
             return (object) [
-                'data' => $response->json(),
+                'data' => [
+                    'success' => false,
+                    'message' => 'Não foi possível criar uma nova conexão.',
+                ],
+            ];
+        } catch (\Exception $exception) {
+            throw new \Exception($exception->getMessage());
+        }
+    }
+
+    public function connect(int|string $connection): array|object
+    {
+        Log::debug(__CLASS__.'.'.__FUNCTION__.' => start', [
+            'connection' => $connection,
+        ]);
+
+        try {
+            if (
+                $this->connectionRepository->exists(column: 'token', value: $connection) &&
+                $this->connectionRepository->exists(column: 'is_active', value: 0)
+            ) {
+                $response = $this->request->get("{$this->url}/instance/connect/{$connection}");
+
+                if ($response->successful()) {
+                    $response = $response->json();
+
+                    return (object) [
+                        'data' => [
+                            'success' => true,
+                            'message' => 'Conexão retornada com sucesso.',
+                            'payload' => (object) $response,
+                        ],
+                    ];
+                }
+            }
+
+            return (object) [
+                'data' => [
+                    'success' => false,
+                    'message' => 'Não foi possivel retornar essa conexão.',
+                ],
             ];
 
         } catch (\Exception $exception) {
@@ -83,6 +169,15 @@ class WhatsappProvinder implements MessengerServiceInterface
         };
 
         $response = $this->request->post("{$this->url}/message/{$endpoint}/{$data['connection']}", $payload);
+
+        $connection = $this->connectionRepository->exists(column: 'is_active', value: 0);
+
+        $createMessage = $this->messageRepository->create([
+            'user_id' => rand(1, 5),
+
+            'is_active' => 0,
+            'payload' => json_encode($response),
+        ]);
 
         return (object) [
             'data' => $response->json() ?? [],
@@ -143,19 +238,6 @@ class WhatsappProvinder implements MessengerServiceInterface
         return $message;
     }
 
-    public function connect(int|string $connection): array|object
-    {
-        Log::debug(__CLASS__.'.'.__FUNCTION__.' => start', [
-            'connection' => $connection,
-        ]);
-
-        $response = $this->request->get("{$this->url}/instance/connect/{$connection}");
-
-        return (object) [
-            'data' => $response->json() ?? [],
-        ];
-    }
-
     public function fetch(string|int $connection): array|object
     {
         Log::debug(__CLASS__.'.'.__FUNCTION__.' => start', [
@@ -174,11 +256,38 @@ class WhatsappProvinder implements MessengerServiceInterface
             'connection' => $connection,
         ]);
 
-        $response = $this->request->delete("{$this->url}/instance/connectionState/{$connection}");
+        try {
+            if (
+                $this->connectionRepository->exists(column: 'token', value: $connection) &&
+                $this->connectionRepository->exists(column: 'is_active', value: 1)
+            ) {
 
-        return (object) [
-            'data' => $response->json() ?? [],
-        ];
+                $response = $this->request->delete("{$this->url}/instance/connectionState/{$connection}");
+
+                if ($response->successful()) {
+
+                    $response = $response->json();
+
+                    return (object) [
+                        'data' => [
+                            'success' => true,
+                            'message' => 'Status da conexão retornado com sucesso.',
+                            'payload' => (object) $response,
+                        ],
+                    ];
+                }
+            }
+
+            return (object) [
+                'data' => [
+                    'success' => false,
+                    'message' => 'Não foi possivel retornar essa conexão.',
+                ],
+            ];
+
+        } catch (\Exception $exception) {
+            throw new \Exception($exception->getMessage());
+        }
     }
 
     public function disconnect(string|int $connection): array|object
@@ -187,11 +296,38 @@ class WhatsappProvinder implements MessengerServiceInterface
             'connection' => $connection,
         ]);
 
-        $response = $this->request->delete("{$this->url}/instance/logout/{$connection}");
+        try {
+            if (
+                $this->connectionRepository->exists(column: 'token', value: $connection) &&
+                $this->connectionRepository->exists(column: 'is_active', value: 1)
+            ) {
 
-        return (object) [
-            'data' => $response->json() ?? [],
-        ];
+                $response = $this->request->delete("{$this->url}/instance/logout/{$connection}");
+
+                if ($response->successful()) {
+
+                    $response = $response->json();
+
+                    return (object) [
+                        'data' => [
+                            'success' => true,
+                            'message' => 'Conexão desconectada com sucesso.',
+                            'payload' => (object) $response,
+                        ],
+                    ];
+                }
+            }
+
+            return (object) [
+                'data' => [
+                    'success' => false,
+                    'message' => 'Não foi possivel desconectar essa conexão.',
+                ],
+            ];
+
+        } catch (\Exception $exception) {
+            throw new \Exception($exception->getMessage());
+        }
     }
 
     public function delete(string|int $connection): array|object
@@ -200,11 +336,37 @@ class WhatsappProvinder implements MessengerServiceInterface
             'connection' => $connection,
         ]);
 
-        $response = $this->request->delete("{$this->url}/instance/delete/{$connection}");
+        try {
+            if (
+                $this->connectionRepository->exists(column: 'token', value: $connection) &&
+                $this->connectionRepository->exists(column: 'is_active', value: 1)
+            ) {
+                $response = $this->request->delete("{$this->url}/instance/delete/{$connection}");
 
-        return (object) [
-            'data' => $response->json() ?? [],
-        ];
+                if ($response->successful()) {
+
+                    $response = $response->json();
+
+                    return (object) [
+                        'data' => [
+                            'success' => true,
+                            'message' => 'Conexão desconectada com sucesso.',
+                            'payload' => (object) $response,
+                        ],
+                    ];
+                }
+            }
+
+            return (object) [
+                'data' => [
+                    'success' => false,
+                    'message' => 'Não foi possivel desconectar essa conexão.',
+                ],
+            ];
+
+        } catch (\Exception $exception) {
+            throw new \Exception($exception->getMessage());
+        }
     }
 
     // Function to handle the webhook
@@ -214,35 +376,41 @@ class WhatsappProvinder implements MessengerServiceInterface
             'data' => $data,
         ]);
 
-        if (isset($data['event'])) {
-            if ($data['event'] === 'QRCODE_UPDATED') {
-                // Cria conexao
-                Log::debug(__CLASS__.'.'.__FUNCTION__.' => QRCODE_UPDATED', [
-                    'data' => $data,
-                ]);
-            }
+        $event = Arr::get($data, 'data.event', false);
 
-            if ($data['event'] === 'MESSAGES_UPSERT') {
-
-                Log::debug(__CLASS__.'.'.__FUNCTION__.' => MESSAGES_UPSERT', [
-                    'data' => $data,
-                ]);
-
-                $this->trigger($data);
-            }
-        }
+        match ($event) {
+            'qrcode.updated' => $this->triggerQrcode($data),
+            'messages.upsert' => $this->triggerFlow($data),
+            default => false,
+        };
 
         return (object) [
             'data' => $data,
         ];
     }
 
-    public function trigger(string|object $data): void
+    public function triggerQrcode(array|object $data): array|object
     {
-        Log::debug(__CLASS__.'.'.__FUNCTION__.' => start', [
+        Log::debug(__CLASS__.'.'.__FUNCTION__.' => QRCODE_UPDATED TRIGGER', [
+            'data' => $data,
+        ]);
+        // atualiza status da conexao is_active com base no callback caso o usuario desconecte pelo telefone.
+
+        return [
+            'data' => $data,
+        ];
+    }
+
+    public function triggerFlow(array|object $data): array|object
+    {
+        Log::debug(__CLASS__.'.'.__FUNCTION__.' => MESSAGES_UPSERT', [
             'data' => $data,
         ]);
 
+        // Recebe a mensagem enviada pelo usuario
 
+        return [
+            'data' => $data,
+        ];
     }
 }
